Refactor Card and Project JavaScript to include Props and be more extendable
Cards can know hold as many logos and tech images as required, using props to display these and .map and .filter to make sure they are correctly displayed if they exist. Furthermore description is now more readible and extendable inside of the code for future adaptations to the projects page, the refactor makes the code more readable and extendable.